Face Emotion Recogniation Using CNN/requirements.txt
streamlit
opencv-python
tensorflow
google-generativeai
python-dotenv
numpy
pillow